better_duck_core::config

Struct Config

pub struct Config { /* private fields */ }

duckdb configuration Refer to https://github.com/duckdb/duckdb/blob/master/src/main/config.cpp

Implementations

impl Config

pub fn enable_autoload_extension(self, enabled: bool) -> Result<Config>

enable autoload extensions

pub fn access_mode(self, mode: AccessMode) -> Result<Config>

Access mode of the database (AUTOMATIC, READ_ONLY, or READ_WRITE)

pub fn custom_user_agent(self, custom_user_agent: &str) -> Result<Config>

Metadata from DuckDB callers

pub fn default_order(self, order: DefaultOrder) -> Result<Config>

The order type used when none is specified (ASC or DESC)

pub fn default_null_order(self, null_order: DefaultNullOrder) -> Result<Config>

Null ordering used when none is specified (NULLS_FIRST or NULLS_LAST)

pub fn enable_external_access(self, enabled: bool) -> Result<Config>

Allow the database to access external state (through e.g. COPY TO/FROM, CSV readers, pandas replacement scans, etc)

pub fn enable_object_cache(self, enabled: bool) -> Result<Config>

Whether or not object cache is used to cache e.g. Parquet metadata

pub fn allow_unsigned_extensions(self) -> Result<Config>

Allow to load third-party duckdb extensions.

pub fn max_memory(self, memory: &str) -> Result<Config>

The maximum memory of the system (e.g. 1GB)

pub fn threads(self, thread_num: i64) -> Result<Config>

The number of total threads used by the system

pub fn with( self, key: impl AsRef<str>, value: impl AsRef<str>, ) -> Result<Config>

Add any setting to the config. DuckDB will return an error if the setting is unknown or otherwise invalid.